aplicado paginacao dos post's, removido arquivo tosco, modificado sistema de tags

+++ contas/views.py
@@ -1,29 +1,37 @@
 # Create your views here.
 from django.shortcuts import render_to_response, get_object_or_404
 from django.template import RequestContext
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.utils.translation import ugettext as _
 from django.core.paginator import Paginator
 from models import ContaPagar, ContaReceber, CONTA_STATUS_APAGAR
 from forms import FormPagamento
+from django.contrib.auth.decorators import login_required
 import datetime
 
+@login_required
 def contas(request):
     contas_a_pagar = ContaPagar.objects.filter(
-        status=CONTA_STATUS_APAGAR,
+        status=CONTA_STATUS_APAGAR,usuario=request.user,
     )
     contas_a_receber = ContaReceber.objects.filter(
-        status=CONTA_STATUS_APAGAR,
+        status=CONTA_STATUS_APAGAR,usuario=request.user,
     )
     return render_to_response('contas/contas.html',locals(), context_instance=RequestContext(request),  )
 
+@login_required
 def conta(request, conta_id, classe):
     conta = get_object_or_404(classe, id=conta_id)
+    if conta.usuario != request.user:
+        raise Http404
     form_pagamento = FormPagamento()
     return render_to_response( 'contas/conta.html', locals(), context_instance=RequestContext(request),  )
 
+@login_required
 def conta_pagamento(request, conta_id, classe):
     conta = get_object_or_404(classe, id=conta_id)
+    if conta.usuario != request.user:
+        raise Http404
     if request.method == 'POST':
         form_pagamento = FormPagamento(request.POST)
       
@@ -32,9 +40,9 @@ def conta_pagamento(request, conta_id, classe):
     
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
 
-
+@login_required
 def contas_por_classe(request, classe, titulo):
-    contas = classe.objects.order_by('status','data_vencimento')
+    contas = classe.objects.filter(usuario=request.user).order_by('status','data_vencimento')
     paginacao = Paginator(contas, 2)
     pagina = paginacao.page(request.GET.get('pagina',1))
     titulo = _(titulo)
@@ -44,10 +52,12 @@ def contas_por_classe(request, classe, titulo):
 #    form = classe_form()
 
 #    return render_to_response( 'contas/editar_conta.html', locals(), context_instance= RequestContext(request),)
-
+@login_required
 def editar_conta(request, classe_form, titulo, conta_id=None):
     if conta_id:
         conta = get_object_or_404( classe_form._meta.model,id=conta_id )
+        if conta.usuario != request.user:
+            raise Http404
     else:
         conta = None
     
@@ -64,9 +74,11 @@ def editar_conta(request, classe_form, titulo, conta_id=None):
 
     return render_to_response( 'contas/editar_conta.html', locals(),context_instance=RequestContext(request), )
 
-
+@login_required
 def excluir_conta(request, classe, conta_id, proxima='/contas/'):
     conta = get_object_or_404(classe, id=conta_id)
+    if conta.usuario != request.user:
+        raise Http404
     conta.delete()
     request.user.message_set.create(
         message='Conta excluida com sucesso!'
